requirements.txt
Keras_Applications==1.0.8
matplotlib==3.6.3
numpy==1.24.1
Pillow==9.4.0
python_telegram_bot==20.0
python_telegram_bot[job-queue]==20.0
tensorflow==2.11.0
tqdm==4.64.1
